# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CRMS_Swagger information",
        "title": "CRMS_Swagger",
        "termsOfService": "http://www.google.com",
        "contact": {
            "name": "Jason Yang",
            "url": "http://www.google.com",
            "email": "jjkk900925@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api",
    "paths": {
        "/customer": {
            "get": {
                "description": "Get Customer by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "example": "L123546789",
                        "description": "Customer ID",
                        "name": "ID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Customer"
                        }
                    },
                    "500": {
                        "description": "{\"Message\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Modify Customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "parameters": [
                    {
                        "description": "Customer Information",
                        "name": "Customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CustomerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Customer"
                        }
                    },
                    "500": {
                        "description": "{\"Message\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "parameters": [
                    {
                        "description": "Customer Information",
                        "name": "Customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CustomerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Customer"
                        }
                    },
                    "500": {
                        "description": "{\"Message\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Customer by CustomerId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer id",
                        "name": "CustomerId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message\": \"Delete success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Message\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customerList": {
            "get": {
                "description": "Get all Customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Customer"
                        }
                    },
                    "500": {
                        "description": "{\"Message\": \"Internal Error!\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/history": {
            "get": {
                "description": "Get History by CustomerId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer id",
                        "name": "CustomerId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.History"
                        }
                    },
                    "500": {
                        "description": "{\"Message\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Modify History",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "parameters": [
                    {
                        "description": "History Information",
                        "name": "History",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HistoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.History"
                        }
                    },
                    "500": {
                        "description": "{\"Message\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new History",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "parameters": [
                    {
                        "description": "History Information",
                        "name": "History",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HistoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.History"
                        }
                    },
                    "500": {
                        "description": "{\"Message\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete History by HistoryId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "History id",
                        "name": "HistoryId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message\": \"Delete success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Message\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/historyForDuring": {
            "post": {
                "description": "Get History For During",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "parameters": [
                    {
                        "description": "History Information",
                        "name": "History",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DuringRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.History"
                        }
                    },
                    "500": {
                        "description": "{\"Message\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/historyList": {
            "get": {
                "description": "Get all History",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.History"
                        }
                    },
                    "500": {
                        "description": "{\"Message\": \"Internal Error!\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Customer": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string"
                },
                "Birthday": {
                    "type": "string"
                },
                "CarNumber": {
                    "type": "string"
                },
                "Citizenship": {
                    "type": "string"
                },
                "CustomerId": {
                    "type": "integer"
                },
                "Gender": {
                    "type": "string"
                },
                "ID": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Note": {
                    "type": "string"
                },
                "PhoneNumber": {
                    "type": "string"
                },
                "history": {
                    "$ref": "#/definitions/model.History"
                }
            }
        },
        "model.CustomerRequest": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string"
                },
                "Birthday": {
                    "type": "string"
                },
                "CarNumber": {
                    "type": "string"
                },
                "Citizenship": {
                    "type": "string"
                },
                "CustomerId": {
                    "type": "integer"
                },
                "Gender": {
                    "type": "string"
                },
                "ID": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Note": {
                    "type": "string"
                },
                "PhoneNumber": {
                    "type": "string"
                }
            }
        },
        "model.DuringRequest": {
            "type": "object",
            "properties": {
                "endDate": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                }
            }
        },
        "model.History": {
            "type": "object",
            "properties": {
                "CustomerId": {
                    "type": "integer"
                },
                "Date": {
                    "type": "string"
                },
                "HistoryId": {
                    "type": "integer"
                },
                "Note": {
                    "type": "string"
                },
                "NumberOfPeople": {
                    "type": "integer"
                },
                "Price": {
                    "type": "integer"
                },
                "Room": {
                    "type": "string"
                }
            }
        },
        "model.HistoryRequest": {
            "type": "object",
            "properties": {
                "CustomerId": {
                    "type": "integer"
                },
                "Date": {
                    "type": "string"
                },
                "HistoryId": {
                    "type": "integer"
                },
                "Note": {
                    "type": "string"
                },
                "NumberOfPeople": {
                    "type": "integer"
                },
                "Price": {
                    "type": "integer"
                },
                "Room": {
                    "type": "string"
                }
            }
        }
    }
}